% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{generateDataExample3}
\alias{generateDataExample3}
\title{Simulate data; Example 3 from Jiang et al. 2019.}
\usage{
generateDataExample3()
}
\value{
List object with the following attributes:
x (proportion of X in each geographic unit);
t (proportion of T in each geographic unit);
n (population of each geographic unit);
bd (the true district B);
}
\description{
\code{generateDataExample3} generates simulated data for example simulation 3 in Jiang et al. 2019.
}
\examples{
generatedData <- generateDataExample3()

}
