% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{generateDataExample5}
\alias{generateDataExample5}
\title{Simulate data; Example 5 from Jiang et al. 2019.}
\usage{
generateDataExample5()
}
\value{
List object with the following attributes:
x (proportion of X in each geographic unit);
t (proportion of T in each geographic unit);
n (population of each geographic unit);
bd (the true district B);
}
\description{
\code{generateDataExample5} generates simulated data for example simulation 5 in Jiang et al. 2019.
}
\examples{
generatedData <- generateDataExample5()

}
