% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{runExampleSimulations}
\alias{runExampleSimulations}
\title{Run illustrative simulations.}
\usage{
runExampleSimulations()
}
\value{
No explicit return values. The summary of the simulations will be printed to standard out.
}
\description{
\code{runExampleSimulations} generates bounds for the example simulations in Jiang et al. 2019.
}
\examples{
runExampleSimulations()

}
